Work in progress to add date pickers.
Committing to save progress. As part of this the PHP backend is being changed to return ISO 8601 formatted dates so that the angular code and javascript code can more easily handle the dates. The date picker class requires that the fields be actual date objects and the angular date filter needs ISO 8601 formatted strings, unix timestamps, or a date object.

[ci skip]