Fix upsert_username action and mysql mode problem.
The upser_username wouldn't correctly convert booleans to ints for storage into the database. With version 5.7 of mysql the sql mode `ONLY_FULL_GROUP_BY` is enabled by default. When this mode is enabled the get bans query doesn't work. This query should be re-written to fix this issue, until then though a query is sent in to disable the ONLY_FULL_GROUP_BY mode.